never mind

+++ source/includes/fundamentals/code-examples/linq.cs
@@ -26,7 +26,7 @@ public class Address
     public string Building { get; set; }
 
     [BsonElement("coord")]
-    public double[] Coordinates { get; set; }
+    public float[] Coordinates { get; set; }
 
     public string Street { get; set; }
